+++ src/settings.ts
@@ -2,7 +2,10 @@ import {TemplateSrv} from "grafana/app/features/templating/template_srv";
 
 export default class SmartSettings {
 
-    constructor(private model: Object, private templateSrv: TemplateSrv) {
+    private _request: Object;
+
+    constructor(private model: Object, private templateSrv: TemplateSrv, private request?: Object) {
+        this._request = this.request || {};
         this.establishProperties();
     }
 
@@ -18,6 +21,20 @@ export default class SmartSettings {
     interpolateVariable(name, variables?: Object) {
 
         variables = variables || {};
+
+        // Put all request parameters from url into the variable
+        // interpolation dictionary, prefixed by `request_`.
+        //
+        // This enables to use request variables in all panel control options.
+        // So, when invoking the dashboard with an url query parameter like
+        // `map_center_latitude=42.42`, you would be able to interpolate it to
+        // a panel control options by i.e. assigning
+        // `mapCenterLatitude: $request_map_center_latitude`.
+        for (let key in this._request) {
+            const value = this._request[key];
+            key = 'request_' + key;
+            variables[key] = value;
+        }
 
         // By default, use vanilla control option attribute from panel data model.
         let value = this.model[name];